GAL_003-1 changed names from main to other names [RF,NE]
