+++ morning07/app/src/main/java/at/tugraz/morning07/MainActivity.java
@@ -21,23 +21,5 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        this.shareButton = this.findViewById(R.id.shareButton);
-        OnClickListenerShare shareListener = new OnClickListenerShare();
-        ArrayList<Uri> imageUris = new ArrayList<>();
-        File f = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS);
-        imageUris.add(Uri.parse(f.getAbsolutePath() + "/test.jpg"));
-        shareListener.setImageArray(imageUris);
-        this.shareButton.setOnClickListener(shareListener);
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        ImageView bigView = (ImageView) findViewById(R.id.big_image);
-        Bundle extras = getIntent().getExtras();
-        byte[] byteArray = extras.getByteArray("bitmap");
-        Bitmap bmp = BitmapFactory.decodeByteArray(byteArray, 0, byteArray.length);
-        bigView.setImageBitmap(bmp);
     }
 }
